requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
discord.py==1.7.1
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
Pillow==8.2.0
pylint==2.6.0
pypng==0.0.20
PyQRCode==1.2.1
pyshorteners==1.0.1
requests==2.25.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3